requirements.txt
flask
firebase_admin
langchain
langchain-google-genai
python-dotenv
requests
flask-cors
faiss-cpu
sentence-transformers
numpy
scikit-learn
pandas
matplotlib
seaborn
plotly
dash
streamlit
fpdf2
reportlab
pymupdf
langchain_community
